refactor: default 메소드를 이용하여 DefaultPathValidator 내부 검증 로직 인터페이스로 이동
